early-boot-config: Add `from_toml_str` to SettingsJson
This adds a method to SettingsJson that takes in a string (expected to
be TOML data), and attempts to parse it, remove the "settings" layer,
and create a `SettingsJson` from it.  This logic was previously
sprinkled around in a few places, creating multiple identical error
variants, etc.
